Report knockout claim fee reward amount in WithdrawKnockout events
While it is possible to derive the portion of a proven knockout claim's
baseFlow and quoteFlow which came from fees, it is quite hard to do so
without perfect context. Instead of reporting that the WithdrawKnockout
event was proven, we now report the amount of the baseFlow/quoteFlow
which comes from fees (0 for no fees or forfeiture).