Redefine incentives contracts
This modifies the structure of the incentives contract we have been
working with to solve a key problem. Namely the accuracy was either
unacceptable or the management of the contract (funding expectations)
where intractable.

This has now been resolved with a new delegated claiming model that can
be added to the existing althea l1 gassless tx ecosystem. This provides
consistent updates that and allows for efficient operation and
management in one package.