Fix tests running in parallel
This patch fixes a number of test ordering issues and import issues to
allow parallel operation.

This only reduces runtime from 3 minutes 38 seconds to 1 minute 50
seconds on my machine but these improvements are still significant to
fix intermittent failures due to poor isolation between the tests
previously and a few test filse that simply would not run in isolation
because the imports where incorrect.